Made the header responsive
Changed the old navbar to a new one that collapses on smartphones to a menu toggle. This required heavy changes to structure of header. The current structure is much simpler. It was taken from Bootstrap's documentation on navigation bars.

Also: Included link to Bootstrap JavaScript in the <head>. This was required for the menu toggle to work.